Issue #7 centralize handling of test suite run id
