+++ .github/workflows/benchmarks.yaml
@@ -30,9 +30,9 @@ jobs:
             --html report/report.html --self-contained-html \
             --track-metrics-json=report/metrics.json \
             --track-metrics-parquet-s3-bucket="APEx-benchmarks" \
-            --track-metrics-parquet-s3-key="tmp/metrics-v4.parquet" \
+            --track-metrics-parquet-s3-key="tmp/metrics-v5.parquet" \
+            --track-metrics-parquet-partitioning="YYYYMM" \
             --basetemp=tmp_path_root \
-            --upload-assets-run-id="gh-$GITHUB_RUN_ID" \
             --upload-assets-s3-bucket="APEx-benchmarks"
         env:
           OPENEO_AUTH_METHOD: client_credentials
